Move colorblind theme to its own file
